fix(issue-quality): accept equivalent structured bug evidence
